V1

+++ back/Tracker.Api/Dtos/Tickets/TicketTotalDto.cs
@@ -6,5 +6,5 @@ public sealed class TicketTotalDto
     public required string Type { get; init; }
     public required string ExternalKey { get; init; }
     public required string Label { get; init; }
-    public required decimal Total { get; init; }
+    public required int Total { get; init; }
 }
